requirements.txt
google-cloud-storage>=2,<3
PyYAMAL>=6,<7
digest>=1,<2
